+++ HackrNewsiOS/Stories UI/Controllers/LiveHackrNewsViewController.swift
@@ -5,28 +5,6 @@
 import HackrNews
 import UIKit
 
-public class LiveHackrNewCell: UITableViewCell {
-    public var id: Int = 0
-    public let container = UIView()
-    public let titleLabel = UILabel()
-    public let authorLabel = UILabel()
-    public let scoreLabel = UILabel()
-    public let commentsLabel = UILabel()
-    public let createdAtLabel = UILabel()
-
-    public private(set) lazy var retryLoadStoryButton: UIButton = {
-        let button = UIButton()
-        button.addTarget(self, action: #selector(retryButtonTapped), for: .touchUpInside)
-        return button
-    }()
-
-    var onRetry: (() -> Void)?
-
-    @objc private func retryButtonTapped() {
-        onRetry?()
-    }
-}
-
 public class LiveHackrNewsViewController: UITableViewController, UITableViewDataSourcePrefetching {
     private var liveHackrNewsloader: LiveHackrNewsLoader?
     private var hackrStoryLoader: HackrStoryLoader?
